IOS-2652 Extract send error to a file
