+++ Tangem/Modules/Send/SendViewModel.swift
@@ -762,26 +762,3 @@ extension SendViewModel {
         }
     }
 }
-
-// MARK: - SendError
-
-struct SendError: Error, BindableError {
-    private let error: Error
-    private let openMailAction: (Error) -> Void
-
-    init(_ error: Error, openMailAction: @escaping (Error) -> Void) {
-        self.error = error
-        self.openMailAction = openMailAction
-    }
-
-    var alertBinder: AlertBinder {
-        let errorDescription = String(error.localizedDescription.dropTrailingPeriod)
-
-        let alert = Alert(title: Text(Localization.feedbackSubjectTxFailed),
-                          message: Text(Localization.alertFailedToSendTransactionMessage(errorDescription)),
-                          primaryButton: .default(Text(Localization.alertButtonRequestSupport), action: { openMailAction(error) }),
-                          secondaryButton: .default(Text(Localization.commonCancel)))
-
-        return AlertBinder(alert: alert)
-    }
-}
